+++ support-frontend/assets/components/marketingConsent/marketingConsent.scss
@@ -1,60 +1,18 @@
-.component-page-section--marketing-consent {
+// -----  gu-sass ----- //
+@import '~stylesheets/gu-sass/gu-sass';
 
-  .component-page-section__heading {
-    font-family: $gu-headline;
-    font-size: 24px;
-    line-height: 1.17;
-    font-weight: bold;
-
-    @include dropline;
-
-    @include mq($until: phablet) {
-      width: 70%;
-    }
-
-    @include mq($from: phablet, $until: desktop) {
-      width: 40%;
-    }
-  }
-
-  .component-page-section__body {
-    padding-top: $gu-v-spacing / 2;
-    padding-bottom: $gu-v-spacing * 2;
-
-    @include mq($from: desktop) {
-      padding-left: $gu-h-spacing;
-    }
-    max-width: 500px;
-  }
-
-  .component-checkbox__title {
-    font-weight: normal;
-    font-size: 16px;
-    margin-bottom: 10px;
-  }
-
-  .component-checkbox__copy {
-    font-family: $gu-text-sans-web;
-  }
-
-  .component-cta-link--next {
-    margin-bottom: $gu-v-spacing * 3;
-    margin-top: $gu-v-spacing * 2;
-
-    @include mq($from: mobileMedium) {
-      width: 300px;
-    }
+.component-marketing-consent {
+  font-weight: 300;
+  margin: 0 -10px;
+  padding: 0 10px ($gu-v-spacing * 2);
+  .svg-newsletters {
+    width: 100%;
   }
 }
 
-.component-marketing-consent__confirmation-message {
-  .component-dotcom-cta {
-    margin-top: $gu-v-spacing * 2;
-  }
-}
-
-.marketing-consent__final-message {
-  font-family: $gu-text-sans-web;
-  font-size: 14px;
-  padding-left: $gu-h-spacing;
+.component-marketing-consent-confirmation {
+  @include gu-fontset-explainer;
+  font-style: italic;
+  margin-top: $gu-v-spacing/4;
+  font-weight: 100;
 }
