--- support-frontend/assets/components/marketingConsentNew/marketingConsent.scss
@@ -1,18 +1,0 @@
-// -----  gu-sass ----- //
-@import '~stylesheets/gu-sass/gu-sass';
-
-.component-marketing-consent {
-  font-weight: 300;
-  margin: 0 -10px;
-  padding: 0 10px ($gu-v-spacing * 2);
-  .svg-newsletters {
-    width: 100%;
-  }
-}
-
-.component-marketing-consent-confirmation {
-  @include gu-fontset-explainer;
-  font-style: italic;
-  margin-top: $gu-v-spacing/4;
-  font-weight: 100;
-}
