+++ support-frontend/assets/pages/contributions-landing/contributionsLanding.scss
@@ -7,8 +7,7 @@
 @import '~components/ctaLink/ctaLink';
 @import '~components/generalErrorMessage/generalErrorMessage';
 @import '~components/button/button';
-@import '~components/marketingConsentNew/marketingConsent';
-@import '~components/marketingConsentNew/marketingConsent';
+@import '~components/marketingConsent/marketingConsent';
 @import '~components/spreadTheWord/spreadTheWord';
 @import '~components/socialShare/socialShare';
 @import '~pages/contributions-landing/components/ContributionThankYou/ContributionThankYou';
